tp: Split ClockSynchronizer into pure engine and ClockTracker wrapper
Refactor ClockSynchronizer from a template class into a concrete class
with a virtual listener interface, and introduce ClockTracker as a
wrapper that adds trace-time semantics (trace-time clock selection,
remote-machine offsets, metadata writes).

This separates concerns: ClockSynchronizer is now a pure clock-domain
conversion engine (graph, snapshots, BFS pathfinding, cache), while
ClockTracker handles trace-processor-specific policy. TraceTimeState
is extracted as shared state owned by TraceProcessorContext.

Key changes:
- ClockId, Clock, ClockTimestamp, ClockSyncErrorType promoted to
  namespace-level types
- Template listener replaced with ClockSynchronizerListener virtual
  base class
- Method bodies moved from header to clock_synchronizer.cc
- Listener callbacks OnTraceTimeClockIdChanged, OnSetTraceTimeClock,
  IsLocalHost removed (logic inlined into ClockTracker)
