+++ src/trace_redaction/collect_clocks.h
@@ -56,8 +56,7 @@ class CollectClocks : public CollectPrimitive {
       const protos::pbzero::TracePacket::Decoder& packet,
       Context* context) const;
 
-  mutable std::vector<RedactorClockSynchronizer::ClockTimestamp>
-      clock_snapshot_;
+  mutable std::vector<ClockTimestamp> clock_snapshot_;
 };
 
 }  // namespace perfetto::trace_redaction
